adding changes and documentation the functions

+++ src/functions.h
@@ -2,5 +2,5 @@ float sum(float a, float b);
 float subtraction(float a, float b);
 float multiplication(float a, float b);
 float division(float a, float b);
-float factorialA(float a);
-float factorialB(float b);
+int factorialA(int a);
+int factorialB(int b);
